+++ src/util/position.ts
@@ -5,6 +5,11 @@ export function rangeInRange(r: Range, range: Range): boolean {
   return positionInRange(r.start, range) === 0 && positionInRange(r.end, range) === 0
 }
 
+export function equalsRange(r: Range, range: Range): boolean {
+  if (!samePosition(r.start, range.start)) return false
+  return samePosition(r.end, range.end)
+}
+
 export function samePosition(one: Position, two: Position): boolean {
   return one.line === two.line && one.character === two.character
 }